memory access violation due after callback #2
The upload module state machine calls a completion callback, but did not
create a copy of its own shared pointer before doing so.  When the
callback is called, the handler may decide to remove the last shared
pointer, which would cause the state machine to be destructed when the
callback returns.  By copying the shared pointer before executing the
callback, as well as moving the callback to the end of the functions
calling it, the issue should be fixed.